Added tree_maintenance.c for routines for changing the tree.
Changed structure Tree so that Nbucket is a member.

Added PruneTree so that size of tree can be reduced whitho
ut losing all the information in the grid.  This requires
some trash collection to free the points at some later tim
e when all the points within a memory block are out of the
 grid.  Code is in tree_maintenance.c

spead_particles - Started on routine to speard halos in third
dimension to increase efficiency of tree force calculation.