Grouped header files into slsimlib.h so that no other headers are needed in the code files.
Canged to Grid structure in high level routines so that i_tree and s_tree are never needed at the highest level.

Changed NewBranch() to give unique branch->number.

Renamed partition_images2() to partition_images_kist().

PruneTrees() (which is used in TrimGrid()) still causes some
error that onlt shows up later in AddPointsToTree() because
some how the branch->points of some barnches point to a point
that is not in the linked list of the image tree!!!!  I hate
you PruneTree()!!!

RefreshTree() changed to RefreshGrid() and moved.