Overhall of VIS observational properties.
Made an abstract base class for Observation call Obs and derived a class ObsVIS.
Noise is based on SC8 background values taken from test fields.

still need to:
Figure out the oversampling and noise interaction which seems to have been changed improperly through
Observation::setPixelSize().  The noise needs to stay in sink with the pixel size.

There is not yet a real conversion from surface brightness to e- in ObsVIS with makes the poisson calculation
inaccurate.